requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
outcome==1.3.0.post0
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
selenium==4.22.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
